+++ CHANGELOG.rst
@@ -22,6 +22,7 @@ Added
 Fixed
 ^^^^^
 - Fix `DatetimeField` use '__year' report `'int' object has no attribute 'utcoffset'`. (#1575)
+- Fix `bulk_update` when using custom fields. (#1564)
 
 0.20.1
 ------